# src/main/resources/api-spec/product-api-spec.yml
openapi: 3.0.0
info:
  title: Cosmo Cats Intergalactic Marketplace API
  description: API for managing products in the intergalactic marketplace
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1

paths:
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve a list of all available products in the marketplace.
      responses:
        '200':
          description: Successful retrieval of product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request
        '500':
          description: Internal server error

    post:
      tags:
        - Products
      summary: Create a new product
      description: Add a new product to the marketplace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid product data
        '500':
          description: Internal server error

  /products/{id}:
    get:
      tags:
        - Products
      summary: Get a product by ID
      description: Retrieve a specific product by its unique identifier.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the product to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '500':
          description: Internal server error

    put:
      tags:
        - Products
      summary: Update a product
      description: Update the details of an existing product in the marketplace.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the product to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '400':
          description: Invalid product data
        '500':
          description: Internal server error

    delete:
      tags:
        - Products
      summary: Delete a product
      description: Remove a product from the marketplace.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the product to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product successfully deleted
        '404':
          description: Product not found
        '500':
          description: Internal server error

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Laser Sword"
        price:
          type: number
          format: float
          example: 499.99
        description:
          type: string
          example: "A high-energy plasma blade."
        categoryId:
          type: string
          format: uuid
          example: "c1f1e2d3-4b5c-678d-9e0f-123456789abc"
        planetOrigin:
          type: string
          example: "Mars"
        rarityLevel:
          type: string
          example: "Legendary"
        isInStock:
          type: boolean
          example: true

    ProductRequest:
      type: object
      required:
        - name
        - price
        - categoryId
        - rarityLevel
      properties:
        name:
          type: string
          example: "Photon Blaster"
        price:
          type: number
          format: float
          example: 299.99
        description:
          type: string
          example: "A compact and powerful energy blaster."
        categoryId:
          type: string
          format: uuid
          example: "d4f1e2c3-5b6a-7c8d-9e0f-234567890def"
        planetOrigin:
          type: string
          example: "Venus"
        rarityLevel:
          type: string
          example: "Rare"
        isInStock:
          type: boolean
          example: false